Disable linux TPM12 support if build without cgo
